# webserver/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/user/new-user":{"post":{"tags":["my-server"],"operationId":"userNew","parameters":[{"name":"family_id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"family_password","in":"query","required":true,"schema":{"type":"string"}},{"name":"family_name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Family"}}}}}}},"/user/get-user":{"post":{"tags":["my-server"],"operationId":"userGet","parameters":[{"name":"family_id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"family_password","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Family"}}}}}}},"/user/change-user":{"post":{"tags":["my-server"],"operationId":"userChange","parameters":[{"name":"family_id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"family_password","in":"query","required":true,"schema":{"type":"string"}},{"name":"new_family_password","in":"query","required":false,"schema":{"type":"string"}},{"name":"new_family_name","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Family"}}}}}}},"/family/new-family":{"post":{"tags":["my-server"],"operationId":"familyNew","parameters":[{"name":"family_id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"family_password","in":"query","required":true,"schema":{"type":"string"}},{"name":"family_name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Family"}}}}}}},"/family/get-family":{"post":{"tags":["my-server"],"operationId":"familyGet","parameters":[{"name":"family_id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"family_password","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Family"}}}}}}},"/family/change-family":{"post":{"tags":["my-server"],"operationId":"familyChange","parameters":[{"name":"family_id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"family_password","in":"query","required":true,"schema":{"type":"string"}},{"name":"new_family_password","in":"query","required":false,"schema":{"type":"string"}},{"name":"new_family_name","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Family"}}}}}}}},"components":{"schemas":{"Family":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"password":{"type":"string"},"name":{"type":"string"}}}}}}